Check if internal edges are present before deleting/ removing them in extrude.

+++ client/js/editor/editor.js
@@ -244,7 +244,7 @@ shapy.editor.Editor.prototype.setCanvas = function(canvas) {
 
   // Initialise the layout.
   this.vp_.width = this.vp_.height = 0;
-  this.layout_ = new shapy.editor.Layout.Double();
+  this.layout_ = new shapy.editor.Layout.Single();
   this.rig(this.rigTranslate_);
 };
